Updated twitter_stock so it separates out the fetching from the api, saving to the database, and analysing the tweets.
Additionally any request to the api will NOT request tweets that are older than tweets we already have AND any retrieved tweet will be saved to the database.

A janky prototype for some code that will fetch and store tweets for the top 100 stock tickers exists in code called "updateHistoricalDatabase". This code just iterates through the top 100 stock tickers and fetches/saves them to the database, hopefully stopping when twitter says we've hit our api limit.